+++ lib/src/models/text_generation/response/sync/text_completion.dart
@@ -1,7 +1,7 @@
 import 'package:yandex_gpt_rest_api/src/models/models.dart';
 
 class TextCompletion {
-  final List<ResultMessage> alternatives;
+  final List<MessageHolder> alternatives;
   final ModelUsage usage;
   final String modelVersion;
 
@@ -19,7 +19,7 @@ class TextCompletion {
   factory TextCompletion.fromJson(Map<String, dynamic> json) {
     return TextCompletion(
       alternatives: List.of(json["alternatives"] as List<dynamic>)
-          .map((i) => ResultMessage.fromJson(i as Map<String, dynamic>))
+          .map((i) => MessageHolder.fromJson(i as Map<String, dynamic>))
           .toList(),
       usage: ModelUsage.fromJson(json["usage"] as Map<String, dynamic>),
       modelVersion: json["modelVersion"] as String,