+++ test/models/text_generation/response/sync/text_completion_test.dart
@@ -19,7 +19,7 @@ void main() {
     });
 
     test("fromJson valid", () {
-      const resMessage = ResultMessage(
+      const resMessage = MessageHolder(
         message: Message.assistant("Hello world!"),
         status: ResultMessageStatus.finalDone,
       );